use u8 rather than bool (~16% faster)

+++ lifegame-core/benches/benchmark.rs
@@ -10,7 +10,7 @@ fn benchmark(c: &mut Criterion) {
         let mut rng = StdRng::seed_from_u64(999);
         let alive_prob = 0.2;
         let data = (0..size * size)
-            .map(|_| rng.gen_bool(alive_prob))
+            .map(|_| rng.gen_bool(alive_prob) as u8)
             .collect::<Vec<_>>();
         group.throughput(Throughput::Bytes(size));
         group.bench_with_input(BenchmarkId::from_parameter(size), &size, |b, &size| {
